refactoring: all map views are in maps controller
