+++ app/controllers/application_controller.rb
@@ -76,6 +76,6 @@ class ApplicationController < ActionController::Base
   # see https://github.com/plataformatec/devise/wiki/How-To:-redirect-to-a-specific-page-on-successful-sign-in
   # https://github.com/plataformatec/devise/wiki/How-To:-redirect-to-a-specific-page-on-successful-sign_in,-sign_out,-and-or-sign_up
   def after_sign_in_path_for(_resource)
-    overview_index_path
+    root_path
   end
 end
