+++ app/controllers/users/registrations_controller.rb
@@ -3,11 +3,11 @@
 module Users
   class RegistrationsController < Devise::RegistrationsController
     def after_sign_up_path_for(resource)
-      overview_index_path(resource)
+      root_path(resource)
     end
 
     def after_update_path_for(resource)
-      overview_index_path(resource)
+      root_path(resource)
     end
 
     # before_action :configure_sign_up_params, only: [:create]
