+++ app/views/application/_logo.html.erb
@@ -1,3 +1,3 @@
-<a class="navbar-brand imi-map-logo" href="<%= overview_index_path %>">
+<a class="navbar-brand imi-map-logo" href="<%= root_path %>">
   <%= image_tag "plane.gif", width: 130, height: 130, alt: "" %>
 </a>
