+++ app/views/application/_navigation_main.html.erb
@@ -1,6 +1,6 @@
 <nav class="navbar navbar-expand-md navbar-inverse " id="imi-maps-navbar-main">
 <!--
-  <a class="navbar-brand imi-map-logo" href="<%= overview_index_path %>">
+  <a class="navbar-brand imi-map-logo" href="<%= root_path %>">
     <%= image_tag "plane.gif", width: 130, height: 130, alt: "" %>
   </a>
 -->
@@ -12,7 +12,7 @@
 
   <div class="collapse navbar-collapse" id="imi-map-navbar-core">
     <ul class="navbar-nav ">
-      <li class="nav-item im-nav-item<%= 'active' if current_page?(overview_index_path) %>"><%= link_to t("header.overview"), overview_index_path, class: 'nav-link im-nav-link imi-map-navlink' %></li>
+      <li class="nav-item im-nav-item<%= 'active' if current_page?(root_path) %>"><%= link_to t("header.overview"), root_path, class: 'nav-link im-nav-link imi-map-navlink' %></li>
       <li class="nav-item im-nav-item<%= 'active' if current_page?(internship_offers_path) %>"><%= link_to t("header.exchange"), internship_offers_path, class: 'nav-link im-nav-link imi-map-navlink' %></li>
       <% if current_user.student? %>
         <li class="nav-item im-nav-item<%= 'active' if current_page?(my_internship_path) %>"><%= link_to t("header.internship"),  my_internship_path, class: 'nav-link im-nav-link imi-map-navlink' %></li>
