+++ app/views/maps/index.html.erb
